requirements_test.txt
meraki==1.40.1
<<<<<<< HEAD
ruff==0.1.6
pip-audit==2.7.2
bandit==1.7.8
pytest-homeassistant-custom-component==0.13.205
playwright==1.39.0
flake8==7.0.0
PyTurboJPEG==1.7.2
orjson>=3.10.7
webrtc-models==0.3.0
filelock==3.20.3
=======
diskcache==5.6.3
aiofiles==23.2.1
setuptools==78.1.1
ruff==0.5.5
pip-audit==2.7.3
bandit==1.7.9
pytest-homeassistant-custom-component==0.13.205
pytest-mock
types-aiofiles==23.2.0.20240106
playwright==1.39.0
pytest-cov
py==1.11.0
tzdata==2024.1
janus==1.0.0
psutil-home-assistant==0.0.1
pillow>=11.0.0
fnv-hash-fast
flake8==7.0.0
webrtc-models==0.3.0
PyTurboJPEG
numpy>=1.26.0
>>>>>>> 44727ea (fix: ci workflow permissions, dependencies and services file)
